+++ db/schemas.js
@@ -5,7 +5,7 @@ mongoose.connect('mongodb://localhost:27017/test');
 
 const db = mongoose.connection;
 db.on('error', console.error.bind(console, 'connection error:')); //eslint-disable-line
-db.once('open', () => console.log('db connected!')); //eslint-disable-line
+// db.once('open', () => console.log('db connected!')); //eslint-disable-line
 
 const { Schema } = mongoose;
 const productSchema = new Schema({
@@ -23,6 +23,8 @@ const productSchema = new Schema({
   ],
 });
 
+const Product = mongoose.model('Product', productSchema, 'products');
+
 const styleSchema = new Schema({
   productId: Number,
   style_id: Number,
@@ -38,9 +40,22 @@ const styleSchema = new Schema({
   ],
 });
 
+const Styles = mongoose.model('Styles', productSchema, 'styles');
+
 const skuSchema = new Schema({
   id: Number,
   style_id: Number,
   size: String,
   quantity: Number,
 });
+
+const Skus = mongoose.model('Skus', productSchema, 'skus');
+
+module.exports = {
+  Product,
+  Styles,
+  Skus,
+  productSchema,
+  styleSchema,
+  skuSchema,
+};
