+++ package.json
@@ -4,7 +4,7 @@
   "description": "A service to provide product information to the Atelier web app",
   "main": "server.js",
   "scripts": {
-    "start": "nodemon server.js",
+    "start": "nodemon server/index.js",
     "test": "jest --collectCoverage",
     "lint": "eslint"
   },
@@ -30,6 +30,8 @@
   "devDependencies": {
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.25.2"
+    "eslint-plugin-import": "^2.25.2",
+    "frisby": "^2.1.3",
+    "supertest": "^6.1.6"
   }
 }
