+++ server/server.js
@@ -1,0 +1,11 @@
+const express = require('express');
+
+const cors = require('cors');
+
+const app = express();
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+app.use(cors());
+
+module.exports = app;
